fix cleanup cron

+++ apps/web/lib/api/partners/delete-partner.ts
@@ -25,15 +25,27 @@ export async function deletePartner({ partnerId }: { partnerId: string }) {
     return;
   }
 
-  const links = partner.programs[0].links;
+  const links = partner.programs.length > 0 ? partner.programs[0].links : [];
 
-  await prisma.customer.deleteMany({
-    where: {
-      linkId: {
-        in: links.map((link) => link.id),
+  if (links.length > 0) {
+    await prisma.customer.deleteMany({
+      where: {
+        linkId: {
+          in: links.map((link) => link.id),
+        },
       },
-    },
-  });
+    });
+
+    await bulkDeleteLinks(links);
+
+    await prisma.link.deleteMany({
+      where: {
+        id: {
+          in: links.map((link) => link.id),
+        },
+      },
+    });
+  }
 
   await prisma.payout.deleteMany({
     where: {
@@ -47,21 +59,11 @@ export async function deletePartner({ partnerId }: { partnerId: string }) {
     },
   });
 
-  await prisma.link.deleteMany({
-    where: {
-      id: {
-        in: links.map((link) => link.id),
-      },
-    },
-  });
-
   await prisma.partner.delete({
     where: {
       id: partner.id,
     },
   });
-
-  await bulkDeleteLinks(links);
 
   if (partner.stripeConnectId) {
     await stripe.accounts.del(partner.stripeConnectId);
